+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated the table creation method in MetaDataStore to improve compatibility across various databases.
 - Replaced JSON data with String format before storage in SQLAlchemy.
 - Implemented storage of byte data in base64 format.
+- Migrated MongoDB Atlas vector search as a standalone searcher like lance.
 
 #### New Features & Functionality
 
@@ -24,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Integrate Monitoring(cadvisor/Prometheus) and Logging (promtail/Loki) with Grafana, in the `testenv`
 - Add `QueryModel` and `SequentialModel` to make chaining searches and models easier.
 - Add `insert_to=<table-or-collection>` to `.predict` to allow single predictions to be saved.
+- Support vLLM (running locally or remotely on a ray cluster)
+- Support LLM service in OpenAI format
+- Add ray compute backend
 
 #### Bug Fixes